module12_MCP/pubmed-mcp-server/pyproject.toml
[project]
name = "pubmed-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "httpx>=0.28.1",
    "mcp[cli]>=1.5.0",
]
